requirements.txt
numpy==2.1.3
pickle-mixin==1.0.2
streamlit==1.36.0
streamlit-option-menu==0.4.0
scikit-learn==1.5.2
google-generativeai==0.8.4
protobuf==4.25.3